requirements.txt
asgiref==3.2.10
autopep8==1.5.4
Django==3.1.2
django-bootstrap-crud-templates==1.1.1
django-bower==5.2.0
django-crispy-forms==1.8.1
django-debug-toolbar==3.2
django-scheduler==0.9.3
icalendar==4.0.7
install==1.3.4
mysqlclient==2.0.1
pkg-resources==0.0.0
pycodestyle==2.6.0
PyMySQL==1.0.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
